backend/foodgram/requirements.txt
Django==3.0.5
django-multiselectfield==0.1.12
djangorestframework-simplejwt==4.4.0
djangorestframework==3.11.0
psycopg2-binary==2.8.6
django-filter==2.4.0
Pillow==9.2.0
python-dotenv==0.20.0
gunicorn==20.1.0
sqlparse==0.4.2
pytz==2022.1
PyJWT==2.3.0
asgiref==3.5.2
djoser==2.1.0
drf-extra-fields==3.4.0
django-import-export==2.1.0
